Move compatibility shims to dedicated file _compat
